+++ api/src/test/java/org/pmiops/workbench/firecloud/FireCloudServiceImplTest.java
@@ -7,8 +7,11 @@ import static org.mockito.Mockito.when;
 
 import com.google.api.client.http.HttpTransport;
 import com.google.cloud.iam.credentials.v1.IamCredentialsClient;
+import com.google.common.cache.LoadingCache;
+import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
 import org.junit.Rule;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
@@ -32,16 +35,24 @@ import org.pmiops.workbench.firecloud.model.FirecloudManagedGroupWithMembers;
 import org.pmiops.workbench.firecloud.model.FirecloudNihStatus;
 import org.pmiops.workbench.firecloud.model.FirecloudSystemStatus;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.beans.factory.config.CustomScopeConfigurer;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Scope;
+import org.springframework.context.support.SimpleThreadScope;
+import org.springframework.web.context.WebApplicationContext;
 
 public class FireCloudServiceImplTest extends SpringTest {
 
   private static final String EMAIL_ADDRESS = "abc@fake-research-aou.org";
+
+  @Autowired
+  @Qualifier(FireCloudCacheConfig.SERVICE_ACCOUNT_REQUEST_SCOPED_GROUP_CACHE)
+  private LoadingCache<String, FirecloudManagedGroupWithMembers> firecloudGroupCache;
 
   @Autowired private FireCloudService service;
 
@@ -49,18 +60,31 @@ public class FireCloudServiceImplTest extends SpringTest {
 
   @MockBean private BillingApi billingApi;
   @MockBean private BillingV2Api billingV2Api;
-  @MockBean private GroupsApi groupsApi;
   @MockBean private HttpTransport httpTransport;
   @MockBean private IamCredentialsClient iamCredentialsClient;
   @MockBean private NihApi nihApi;
   @MockBean private ProfileApi profileApi;
   @MockBean private StatusApi statusApi;
 
+  @MockBean
+  @Qualifier(FireCloudConfig.SERVICE_ACCOUNT_GROUPS_API)
+  private GroupsApi groupsApi;
+
   @Rule public MockitoRule mockitoRule = MockitoJUnit.rule();
 
   @TestConfiguration
-  @Import({FireCloudServiceImpl.class, RetryConfig.class})
+  @Import({FireCloudCacheConfig.class, FireCloudServiceImpl.class, RetryConfig.class})
   static class Configuration {
+    @Bean
+    public CustomScopeConfigurer customScopeConfigurer() {
+      // Configure the request scope so we can reuse the cache config; this is not available by
+      // default in Spring unit tests.
+      CustomScopeConfigurer scopeConfigurer = new CustomScopeConfigurer();
+      scopeConfigurer.setScopes(
+          ImmutableMap.of(WebApplicationContext.SCOPE_REQUEST, new SimpleThreadScope()));
+      return scopeConfigurer;
+    }
+
     @Bean
     @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public WorkbenchConfig getWorkbenchConfig() {
@@ -75,6 +99,11 @@ public class FireCloudServiceImplTest extends SpringTest {
     workbenchConfig.firecloud.debugEndpoints = true;
     workbenchConfig.firecloud.timeoutInSeconds = 20;
     workbenchConfig.billing.accountId = "test-billing-account";
+  }
+
+  @AfterEach
+  public void tearDown() {
+    firecloudGroupCache.invalidateAll();
   }
 
   @Test
